refactor(core): better representation of errors in logs and service menus
